- [Mac] Converted all binaries and prebuilt libraries to x86_64 (min 10.6) only - [Mac] Removed Xcode3 project and Xcode3 build scripts - [Mac] Converted sources and projects to ARC and Objective C 2.0 syntax - [Mac] Silenced most of the compiler and static analyser warnings in OUR code - [Mac] Removed the compatibility module proximus - no longer needed - [Mac] Modernized standalone application full screen behaviour. - [Mac] Converted installer scripts to use pkgbuild + productbuild Apple tools. No more PackageMaker - [Mac] Removed PackageMaker installer source files - [Mac] Flattened installer and signing with "Installer" DeveloperID certificate - [Mac] Updated Sparkle framework to version 1.8 - [Mac] Updated ffmpeg library to version 2.4
- Sealed hopefully all remaining obvious memory leaks
- Updated tinyxml library to 2.6.2

+++ trunk/client_generic/Common/MTimer.h
@@ -59,7 +59,7 @@ class CMTimer : public ITimer
         mach_timebase_info_data_t tTBI;
 
         mach_timebase_info(&tTBI);
-        fp8 cv = ((long double) tTBI.numer) / ((long double) tTBI.denom);
+        fp8 cv = (static_cast<fp8>(tTBI.numer)) / (static_cast<fp8>(tTBI.denom));
 
         uint64_t now = mach_absolute_time();
         fp8 tNS = now * cv;
@@ -96,7 +96,7 @@ class CMTimer : public ITimer
 				mach_timebase_info_data_t tTBI;
 
                 mach_timebase_info(&tTBI);
-                fp8 cv = ((long double) tTBI.numer) / ((long double) tTBI.denom);
+                fp8 cv = (static_cast<fp8>(tTBI.numer)) / (static_cast<fp8>(tTBI.denom));
 
 				return cv * 1e-9;
 			}
